pid 파라미터 조정

+++ User/Src/pid_logic.c
@@ -10,7 +10,7 @@
 
 // PID 계산 함수
 // pid_logic.c 수정 버전
-uint16_t PID_Compute(PID_TypeDef *pid, float current_angle)
+float PID_Compute(PID_TypeDef *pid, float current_angle)
 {
     // 1. 타입을 float 또는 int16_t로 변경하여 음수 오차 허용
     float error = (float)pid->target - current_angle;
@@ -35,10 +35,9 @@ uint16_t PID_Compute(PID_TypeDef *pid, float current_angle)
     // 전체 출력 계산
     float total_output = p_out + i_out + d_out;
 
-    // 서보 모터 기준점(예: 1500)을 더해줘야 할 수도 있음
-    // 여기서는 일단 saturation만 처리
-    if (total_output > pid->out_max) total_output = pid->out_max;
-    else if (total_output < pid->out_min) total_output = pid->out_min;
+    // 여기서 -500 ~ 500 사이로 가둡니다.
+	if (total_output > pid->out_max) total_output = pid->out_max;
+	else if (total_output < pid->out_min) total_output = pid->out_min;
 
-    return (uint16_t)total_output;
+    return total_output;
 }
